last one

+++ src/main/java/com/kodilla/ecommercee/entity/Cart.java
@@ -36,7 +36,7 @@ public class Cart {
 
     @Column(name = "user_id")
     @NotNull
-    private BigInteger userID;
+    private Long userID;
 
 
 }
